create archives page, separate jobs page, and applications page to see all candidate applications
